Updated some confusing wording
In the "Back & Forth" section within the 2 points where it describes how the compiler will proceed. The compiler point contained the sentence "The code Engine runs will first ask..." which sounds and feels  grammatically wrong. I updated the wording to be "When the code Engine runs, it first asks..." which feels much better. I may be reading the original text wrong, but wanted to throw this out there just in case it improved the reading of it.